+++ backend/application/api/src/main/kotlin/io/raemian/api/auth/controller/AuthController.kt
@@ -1,7 +1,7 @@
 package io.raemian.api.auth.controller
 
 import io.raemian.api.auth.controller.request.UpdateUserRequest
-import io.raemian.api.auth.controller.response.UserResponse
+import io.raemian.api.user.controller.response.UserResponse
 import io.raemian.api.auth.domain.CurrentUser
 import io.raemian.api.auth.service.AuthService
 import io.raemian.api.lifemap.LifeMapService
@@ -21,34 +21,4 @@ class AuthController(
     private val lifeMapService: LifeMapService,
 ) {
 
-    @Operation(summary = "토큰 유저 정보 조회 API")
-    @GetMapping("/my")
-    fun my(@AuthenticationPrincipal currentUser: CurrentUser): ResponseEntity<ApiResponse<UserResponse>> {
-        val user = authService.getUserById(currentUser.id)
-        val lifeMap = lifeMapService.findFirstByUserId(currentUser.id)
-        val response = UserResponse.of(user, lifeMap)
-        return ResponseEntity.ok(ApiResponse.success(response))
-    }
-
-    @Operation(summary = "유저 온보딩 이후 정보 업데이트 API")
-    @PutMapping("/my")
-    fun update(
-        @AuthenticationPrincipal currentUser: CurrentUser,
-        @RequestBody updateUserRequest: UpdateUserRequest,
-    ): ResponseEntity<Void> {
-        authService.update(
-            id = currentUser.id,
-            nickname = updateUserRequest.nickname,
-            birth = updateUserRequest.birth,
-            username = updateUserRequest.username,
-        )
-        return ResponseEntity.ok().build()
-    }
-
-    @Operation(summary = "유저 삭제")
-    @DeleteMapping("/my")
-    fun delete(@AuthenticationPrincipal currentUser: CurrentUser): ResponseEntity<Void> {
-        authService.delete(currentUser.id)
-        return ResponseEntity.ok().build()
-    }
 }
